Remove 'welcome' challenge for an initial config
This is quite a change. There was a chicken-and-egg problem:

We wanted to automatically hint the user with the file name to use for
his challenges. During development, this was taken from the extension
chosen from welcome.* but then the name displayed for the 'welcome'
challenge was misleading because there was no welcome.* file by then.

So let's remove the 'welcome' challenge altogether and let all challenges
depend con a configuration file which, if doesn't exist, displays the
help text that was originally in the 'welcome' challenge.

In this initial configuration three directives will be stored: language,
mimetype and extension. Sample configurations will be provided for the
user under the config-samples directory. The user should only need to take
the chosen sample and copy it to the working directory.

The welcome message was moved to test/messages/welcome, which is now a
plain text file.

References to the welcome file have been updated.